+++ mr_dapa/components/components.py
@@ -3,3 +3,5 @@ from .lines import LinesComponent as LinesComponent
 from .map import MapComponent as MapComponent
 from .scatter import ScatterComponent as ScatterComponent
 from .fill import FillComponent as FillComponent
+from .heatmap import HeatmapComponent as HeatmapComponent
+from .map3d import Map3DComponent as Map3DComponent